% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorplot.R
\name{errorplot}
\alias{errorplot}
\title{Plot quantities representing errors in Cornish-Fisher approximation, relative to uncorrected normal theory approximation.}
\usage{
errorplot(out, nn = 1, mytitle = "Exponential Example")
}
\arguments{
\item{out}{a matrix, generally provided by testcorn.  See testcorn documentation for details.}

\item{nn}{the sample size.  If nn is not one, it is presumed that the cumulants of the sampling distribution under investigation behave in teh usual way; that is, third cumulants are k3/sqrt(nn), and fourth cumulants are k4/(nn.}

\item{mytitle}{Title for plot.}
}
\value{
None.
}
\description{
Plot quantities representing errors in Cornish-Fisher approximation, relative to uncorrected normal theory approximation.
}
\examples{
#Generate Fig 1 of Chen, Kolassa, Seifu, and Zhong.  Next line takes 8 hours.  Uncommented example
#is not big enough to be useful, but runs quickly.
# errorplot(testcorn(xv=makebigxv(100),nsamp=500000,nn=1)$tail)
errorplot(testcorn(xv=makebigxv(10),nsamp=5000,nn=1)$tail)
}
