% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialcumulants.R
\name{exponentialcumulants}
\alias{exponentialcumulants}
\title{Calculate cumulants for the linear compbination of independent exponential random variables, for exhibition of bivariate Cornish Fisher expansion.}
\usage{
exponentialcumulants(amat = rbind(c(1, 1, 0), c(0, 1, 1)))
}
\arguments{
\item{amat}{a vector with two rows and as many columns as there are independent exponentials}
}
\value{
A list with compoents k1, k2, k3, and k4, representing first through fourth comulant matrices respectively, having 1 through four dimensions, each dimension 2, respectively.
}
\description{
Calculate cumulants for the linear compbination of independent exponential random variables, for exhibition of bivariate Cornish Fisher expansion.
}
