% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebigxv.R
\name{makebigxv}
\alias{makebigxv}
\title{Calculate a grid of points on which to test a Cornish-Fisher expansion.}
\usage{
makebigxv(nps)
}
\arguments{
\item{nps}{number of points (less one) on for each dimension of the resulting grid.}
}
\value{
Grid of ordinates.
}
\description{
Calculate a grid of points on which to test a Cornish-Fisher expansion.
}
