% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcorn.R
\name{testcorn}
\alias{testcorn}
\title{Calculate quantities involved in a Cornish-Fisher two-dimensional quantile approximation, for use in assessing relative accuracy of this approximation.}
\usage{
testcorn(alpha = c(0.1, 0.05), xv = NULL,
  cumulants = exponentialcumulants, amat = rbind(c(1, 1, 0), c(0, 1,
  1)), randf = rexp, nsamp = 50000, nn = 1)
}
\arguments{
\item{alpha}{a vector with two components giving the first univariate upper target tail probability, and the second giving the bivariate upper tail probability.}

\item{xv}{a vector with two components giving the the ordinates associated with the quantiles; this may alternatively be a matrix with two rows and multiple columns, and the approximation will be evaluated at the pair in each column.  This argument takes precedence over alpha.}

\item{cumulants}{a function that generates a list of bivariate cumulant arrays, labeled k1 through k4, with 1 through 4 dimensions respectively, with each dimension 2.}

\item{amat}{a matrix transforming independent random variables into components of a dependent vector.}

\item{randf}{a function giving random observations, to be combined using amat below, to get a set of random vectors with two compoents.}

\item{nsamp}{The number of samples contributing to the Monte Carlo approximation to tail probabilities and quantiles.}

\item{nn}{the sample size.  If nn is not one, it is presumed that the cumulants of the sampling distribution under investigation behave in teh usual way; that is, third cumulants are k3/sqrt(nn), and fourth cumulants are k4/nn.}
}
\value{
A list with two components:
   An array with three dimensions, the first indicating quantities for the first dimension of the sample space, and its marginal distribution, and the second representing the second component and the joint distribution.  The second represents 1. Target upper probability, 2. The associated bivariate normal quantile, 3. The Edgeworth tail approximation for the normal quantile, 4. The Monte Carlo tail approximation for the normal quantile, 5. The Cornish-Fisher quantile, 6. The Monte Carlo tail approximation for the Cornish-Fisher quantile, 7. mean, 8. Standard deviations, 9. quantiles.
   The Monte Carlo sample drawn for comparison purposes.
}
\description{
Calculate quantities involved in a Cornish-Fisher two-dimensional quantile approximation, for use in assessing relative accuracy of this approximation.
}
